+++ src/main/java/com/maxmind/geoip2/record/CityRecord.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * 
  * This record is returned by all the end points except the Country end point.
  */
-public class CityRecord extends AbstractNamedRecord {
+final public class CityRecord extends AbstractNamedRecord {
     @JsonProperty
     private Integer confidence;