+++ src/main/java/com/maxmind/geoip2/record/LocationRecord.java
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * 
  * This record is returned by all the end points except the Country end point.
  */
-public class LocationRecord {
+final public class LocationRecord {
     @JsonProperty("accuracy_radius")
     private Integer accuracyRadius;