+++ src/main/java/com/maxmind/geoip2/record/PostalRecord.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
  * 
  * This record is returned by all the end points except the Country end point.
  */
-public class PostalRecord {
+final public class PostalRecord {
     @JsonProperty
     private String code;