+++ src/test/java/com/maxmind/geoip2/DatabaseReaderTest.java
@@ -13,7 +13,9 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import com.maxmind.geoip2.exception.AddressNotFoundException;
+import com.maxmind.geoip2.exception.GeoIp2Exception;
 import com.maxmind.geoip2.model.City;
+import com.maxmind.geoip2.model.CityIspOrg;
 import com.maxmind.geoip2.model.Omni;
 
 @SuppressWarnings("static-method")
@@ -23,37 +25,37 @@ public class DatabaseReaderTest {
     public ExpectedException exception = ExpectedException.none();
 
     @Test
-    public void testDefaultLanguage() throws IOException,
-            AddressNotFoundException {
+    public void testDefaultLanguage() throws IOException, GeoIp2Exception {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
-        City city = reader.get(InetAddress.getByName("81.2.69.160"));
+        City city = reader.city(InetAddress.getByName("81.2.69.160"));
         assertEquals("London", city.getCity().getName());
         reader.close();
     }
 
     @Test
-    public void testLanguageList() throws IOException, AddressNotFoundException {
+    public void testLanguageList() throws IOException, GeoIp2Exception {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"), Arrays.asList("xx", "ru",
                 "pt-BR", "es", "en"));
-        Omni city = reader.get(InetAddress.getByName("81.2.69.160"));
+        Omni city = reader.omni(InetAddress.getByName("81.2.69.160"));
         assertEquals("Лондон", city.getCity().getName());
         reader.close();
 
     }
 
     @Test
-    public void hasIpAddress() throws IOException, AddressNotFoundException {
+    public void hasIpAddress() throws IOException, GeoIp2Exception {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
-        City city = reader.get(InetAddress.getByName("81.2.69.160"));
-        assertEquals("81.2.69.160", city.getTraits().getIpAddress());
+        CityIspOrg cio = reader
+                .cityIspOrg(InetAddress.getByName("81.2.69.160"));
+        assertEquals("81.2.69.160", cio.getTraits().getIpAddress());
         reader.close();
     }
 
     @Test
-    public void unknownAddress() throws IOException, AddressNotFoundException {
+    public void unknownAddress() throws IOException, GeoIp2Exception {
         this.exception.expect(AddressNotFoundException.class);
         this.exception
                 .expectMessage(containsString("The address 10.10.10.10 is not in the database."));
@@ -61,7 +63,7 @@ public class DatabaseReaderTest {
         DatabaseReader reader = new DatabaseReader(new File(
                 "test-data/GeoIP2-City.mmdb"));
         @SuppressWarnings("unused")
-        City city = reader.get(InetAddress.getByName("10.10.10.10"));
+        City city = reader.city(InetAddress.getByName("10.10.10.10"));
         reader.close();
     }